+++ lib/Rest.php
@@ -163,7 +163,7 @@ class Rest {
         }
 
         // get the directory.
-        $directory = strtolower( $params['directory'] );
+        $directory = $params['directory'];
 
         // get the login.
         $listing_base_object->set_login( $params['login'] );
@@ -233,7 +233,7 @@ class Rest {
 
         // check if all directories have been loaded.
         $directory_loading = false;
-        foreach ( $directory_list as $url => $entry ) {
+        foreach ( $directory_list as $entry ) {
             // bail if loading is already enabled.
             if ( $directory_loading ) {
                 continue;
@@ -269,10 +269,10 @@ class Rest {
          * @param array $directory_list The list of directories.
          * @param string $directory The directory to load.
          */
-        $directory_loading = apply_filters( Init::get_instance()->get_prefix() . '_service_' . $listing_base_object->get_name() . '_directory_loading', $directory_loading, $directory_list, $directory );
+        $directory_loading = apply_filters( Init::get_instance()->get_prefix() . '_directory_listing_' . $listing_base_object->get_name() . '_directory_loading', $directory_loading, $directory_list, $directory );
 
         // bail if we must load further directories.
-        if ( $directory_loading ) {
+        if ( $directory_loading && ! $params['cancelLoading'] ) {
             return array(
                 'directory_loading' => true,
                 'directory_to_load' => count( $next_directories ),